+++ .github/workflows/pr-test.yml
@@ -28,8 +28,8 @@ jobs:
 
       - name: Install libtraceable-downloader
         run: |
-          cd filters/blocking/cmd/libtraceable-downloader &&
-          go install github.com/Traceableai/goagent/filters/blocking/cmd/libtraceable-downloader
+          cd ./filters/traceable/cmd/libtraceable-downloader &&
+          go install github.com/Traceableai/goagent/filters/traceable/cmd/libtraceable-downloader
 
       - name: Pull libtraceable
         run: $(go env GOPATH)/bin/libtraceable-downloader pull-library
@@ -108,8 +108,8 @@ jobs:
         uses: docker/build-push-action@v2
         with:
           push: false
-          context: ./filters/blocking/cmd/libtraceable-downloader
-          file: ./filters/blocking/cmd/libtraceable-downloader/Dockerfile.${{ steps.split.outputs._0}}.test
+          context: ./filters/traceable/cmd/libtraceable-downloader
+          file: ./filters/traceable/cmd/libtraceable-downloader/Dockerfile.${{ steps.split.outputs._0}}.test
           build-args: |
             ${{ steps.string.outputs.uppercase }}_VERSION=${{ steps.split.outputs._1}}
             GO_VERSION=1.16.8
