+++ README.md
@@ -3,6 +3,6 @@
 ## Installing filter
 
 ```go
-go install github.com/Traceableai/goagent/filters/blocking/cmd/libtraceable-downloader
+go install github.com/Traceableai/goagent/filters/traceable/cmd/libtraceable-downloader
 libtraceable-downloader pull-library
 ```
