+++ _tests/main.go
@@ -4,14 +4,14 @@ import (
 	"context"
 
 	traceconfig "github.com/Traceableai/agent-config/gen/go/v1"
-	"github.com/Traceableai/goagent/filters/blocking"
+	"github.com/Traceableai/goagent/filters/traceable"
 	"github.com/hypertrace/goagent/instrumentation/hypertrace"
 	"github.com/hypertrace/goagent/sdk"
 )
 
 func main() {
 	cfg := traceconfig.Load()
-	f := blocking.NewBlockingFilter(cfg)
+	f := traceable.NewFilter(cfg)
 
 	_, s, ender := hypertrace.StartSpan(context.Background(), "test", &sdk.SpanOptions{})
 	defer ender()
