+++ instrumentation/internal/filter/resolver.go
@@ -2,7 +2,7 @@ package filter
 
 import (
 	traceconfig "github.com/Traceableai/agent-config/gen/go/v1"
-	"github.com/Traceableai/goagent/filters/blocking"
+	"github.com/Traceableai/goagent/filters/traceable"
 	sdkfilter "github.com/hypertrace/goagent/sdk/filter"
 )
 
@@ -17,7 +17,7 @@ func isNoop(f sdkfilter.Filter) bool {
 // ResolveFilter resolves a joint filter based on the agent configuration and a provided filter.
 // If both are nil or noop, this function will return nil.
 func ResolveFilter(cfg *traceconfig.AgentConfig, f sdkfilter.Filter) sdkfilter.Filter {
-	blockingFilter := blocking.NewBlockingFilter(cfg)
+	blockingFilter := traceable.NewFilter(cfg)
 
 	if isNoop(blockingFilter) {
 		return f
